pyproject.toml
[project]
name = "yt-pilot"
version = "0.1.0"
description = "YouTube comment collection tool for thesis pilot study"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "python-dotenv>=0.19.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]